project15_nql_new/mgedata/utils/pagination.py
from collections import OrderedDict

from rest_framework.pagination import PageNumberPagination
from rest_framework.response import Response


class MgePageNumberPagination(PageNumberPagination):
    """
    因为DRF自带的分页器不带page_size参数
    所以使用我们自定义的分页器，同时支持page_size参数控制分页大小
    """
    page_size_query_param = 'page_size'
    max_page_size = 50

    def get_paginated_response(self, data):
        return Response(OrderedDict([
            ('count', self.page.paginator.count),
            ('page_size', self.get_page_size(self.request)),
            ('next', self.get_next_link()),
            ('previous', self.get_previous_link()),
            ('results', data)
        ]))
