project15_nql_new/pyproject.toml
[tool.poetry]
name = "mgedata"
version = "2.0.0"
description = "MGEDATA"
authors = ["Harold Chen <harold7233@hotmail.com>"]

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.0"
mongoengine = "^0.23.0"
itsdangerous = "~=1.1.0"
djangorestframework = "^3.13.1"
django-daterange-filter = "^1.3.0"
django-hardcopy = "^0.1.4"
Pillow = "^8.2.0"
qrcode = "^7.3.1"
django-oauth-toolkit = "^1.2.0"
django-cors-middleware = "^1.3.1"
pycryptodome = "^3.15.0"
pandas = "^1.4.3"
elasticsearch-dsl = "^7.0.0"
elasticsearch = "^7.0.0"
openpyxl = "^3.0.10"
patool = "^1.12"
lxml = "^4.9.1"
python-memcached = "^1.59"
django-proxy = "^1.2.1"
psycopg2-binary = "^2.8.6"
celery = "^4.0.0"
tqdm = "^4.64.1"
uwsgi = { version = "^2.0.21", platform = "linux" }
antlr4-python3-runtime = "4.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"